Break down operators.py into multiple Python modules
